Initial commit to add a data scraper
GetFactors() is meant to be a to a convenience function to download financial data sets from a set of resources (research libraries such as Fama-French's data library and Shiller's data library, financial companies data sets such as those open sourced by AQR).

A common issue with target sources is that they don't have an API to connect to, which would allow systematic and massive data access. Hence there's a need to resort to kind of 'artisanal' methods.
This commit gave some initial structure in that direction and several improvements are WIP.

To go on the web 'XML' is currently used, specifically XML::getHTMLLinks(). I have can provide an alternative solution with 'rvest' if needed.

The function also comes with minimal docs, which will be expanded along the way.