Add FF3 model replication attempts with FactorAnalytics
These first attempts use FactorAnalytics::fitFfm() and the idea is to understand whether it can help with the first step of the Fama-MacBeth regression analysis.

Code is added to the replication example vignette as a chunk under the model to reproduce. I'm slightly in doubt whether to put this sort of snippets in the sandbox first.

See #4