Revise and improve GetFactors()
- fix many aspects to handle and import data in R
- switch inner src='FF' switch to ladder
- add 'country' param, thus change a previous annoying flow
- remove 'verbose' param
- add more documentation, fix some typos
- add more examples with new functionality

Getting data from specific countries/regions is left aside at the moment as not strictly necessary at this stage. However, the flow is already set.

Likewise, the 'verbose' param would be nice to have to retrieve descriptions. But as of now, it'is time consuming to implement.

One thing that would certainly improve the function is a more advanced regex. But foundations are shaking anyways, because as the source renames or moves resources the function likely breaks and this becomes burdersome to maintain.

Lastly, col names in the end aren't strictly necessary. It's just an attempt to homogenize variables we'll use often.

GetFactors() was introduced by 73073b0.

See #7