Add basic Fama-MacBeth snippet to the vignette
This is a basic implementatation of the two main steps in the Fama-MacBeth method, which is used to estimate factor models regression coefficients. Its application in the snippet is with respect to the Fama-French Three-factor model, other examples of models on which has been used are mentioned in #4.

Certainly not the only method available, it has been and still remains a popular method in the literature.

There are aspects currently overlooked, to share the main idea. I am working on more comprehensive solutions.