Add object-level documentation templates
This commit closes #10, but it also provides the basic structure to take advantage of roxygen's (great) templating feature in all the other overlapping documentation. This is expected to have an helpful impact on documentation writing.

On the documentation itself, I've tried not to blindly copy-paste the egregious
work of Prof. K.R. French (and Prof. E.F. Fama?). Sure, most of their documentation
is transplanted here because (a) we are documenting exactly their work and the data sets they publicly share and (b) there's really little to edit on the pages and pages of very detailed documentation they share. However, I took the occasion to be slightly more generous on some points while introducing the variables names we'll use. Hope users will enjoy. Please, double check the material referring to their website. Also, Prof. K. R. French name and links to its website are everywhere, it has to be absolutely clear that he is the copyright holder and deserves all credits.

The param 'verbose' in GetFactors() is not currently used and produces a build WARNING.
'verbose' was mainly meant for the purpose being achieved with the now available
abundant documentation on data sets currently downloadable.
I'm therefore removing this param for now.