Add summaries on AF (2013) main points to vignette
These notes are my summaries on the main points of the paper "The Devil in HML's Details" by Clifford Asness and Andrea Frazzini, 2013. Sooner or later this material needed to be added to the vignette, here it is.

Important for replication purposes: I added many comments, they are meant for your attention. Please, provide your feedback. (Commits can be commented on GitHub as well if you prefer).

#13