Solve data sourcing from parsers to vignettes
By default knitr sets 'PackageRoot/vignettes' as wd while knitting. Redefining vignettes root.dir to package root gives a better position to both access other files and keep them individually valid. This is especially useful in our case, because we need parsers to source in data. MSCI's series are the main reason this solution was chosen: in all other cases so far we're able to write directly to disk thanks to 'rio' (or related packages), while MSCI website is more complex and files rely on dynamically generated links. For this reason, I've explicitly added instructions to proceed with the manual download (already needed before). To avoid the manual step of downloading MSCI files, a more sophisticated approach would be required.

Why aren't we adopting the best practice of loading data sets in one of 'data/', 'inst/extdata', or 'vignettes/data'? Because it's uncertain if we can directly share data series. It's all public data, so everyone can access it freely and our parsers will make that easier, but redistributing this data is apparently another pair of shoes and we certainly don't want to infringe anyone's rights.

As a side note, while working on these files I took a chance to make their paths platform agnostic.

Lastly, vignettes building still fails. It's now perhaps more convincing that is something specific about `source("Value--Devil-in-HMLs-Details.R")`. In view of the present solution, I'm wondering if that file souldn't live in inst/parsers where perhaps it better fits.

#15