Merge plots objects
R is well known also for its great plotting capabilities. More elaborated plots are usually built by layering, but although this aspect is very intuitive with interactive use, it is perhaps not so much to knitr while knitting. Or at least I'm not sure about the xts interaction.

As a simple solution, introduced dummy functions so to "merge" plots results and force them to be displayed once. This solution points to the eventual next step, likely an S3 method.

#15