Fix FF data sets monthly dates as of month-end
FF monthly series dates look to be as the end of month. FF don't include days unfortunately, which is a practice unfeasible to do in R as it fortunately causes malformed "Dates". Hence, for example "... return of all CRSP firms incorporated in the US and listed on the NYSE, AMEX, or NASDAQ that have a CRSP share code of 10 or 11 at the beginning of month..." (https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/f-f_factors.html), apparently only refers to the selection criteria. Although there are componentwise differences in a strict sense, checks against AQR's US factors series seem consistent with the above.

Corrections in #15 followed and, once the present commit is assessed, 'FF3.monthly' data set added in #10 and used in #4 should be revised as well.

While we can potentially use AQR's factor data from #13, at this time it's unclear if this is what authors in #15 actually do. Because the methodology is very similar to FF, if not the same, we don't expect a significant influence overall, but it's a fact that there exist differences with FF data sets and these should be based at least on the assets' universe considered. Whichever series we use in the end would need a motivation, thus all of them are needed.

See also #18.