Extract estimates and stats from fitted models
Models of interest from which estimates and stats can now be extracted are AMP's combos, CAPM, FF3 and FFC4. We are estimating more models than shown in the paper, since dimensions aren't explosive there's little cost to do so and it hopefully results in more concise and clearer code.

As discussed, R is known as statisticians' wonderland and it won't be particularly difficult to extract and display fitting results in a pretty printable version. The convenient ExtractModelStats() is rather meant to help ourselves while developing and have what we need right into the console. Also, it's an effort towards code repetitions reduction and better readability.

#15