Add bond factors data increasing US pricing models
These factors should at least proxy FF's TERM and DEF. Importantly, we now estimate the FF6 model for the US. The original series names are left at the moment. For info on these series and motivation on their use see discussions in same main issue, where authors are cited. Bonds factor data parsing was added with 61b97b3.

As with MSCI data introduced with f5633f3, bonds series are being linearly interpolated as a simple way to account for dates mismatches. Refer to #18 for broader discussions on this topic.

#15