Add knitr::kable to print the tables of betas
I started looking at adding the t-values (to more closely resemble Exhibit 6), but it took too long for Canada Day. You can use the output from summary() which includes t-values, and then something like lapply(dfList, '[[', 2) to subset them from the list of dataframes. But it felt wrong to do this for t-values and not for betas (aka "Estimate") as well, and when trying to update the beta subsetting with this approach i realized it was taking too long. Moving on. T-values in the summary output nevertheless. Might be nice to add to the table at a later stage.