Create "Century of Factor Premia" data parser
This truly is a great data set from AQR, where the company gathers a wealth of series reaching 100+ years worth of data in some instances. A century of data, across asset classes/markets and countries/regions.

Data or construction methodologies also include #13 and #15, providing a nice blending of everything we've been working on so far.

At the time of writing, the most evident use case is importantly #17. Considering GSCI and BOND factors are not readily accessible for people who do not have a Bloomberg or Reuters terminal at their fingertips. Hence the need of, hopefully, highly correlated proxies (proxies of proxies).

Further, it seems ubiquitously helpful for the present project and I think we'll use it in the very next future (e.g., "carry" is included as well).

It's a resource not only for ourselves, but for who'll comes next too.