Fix dates mismatch handling
I believe that in our specific framework the kind of values interpolation in place is uncorrect. An interpolation on a monthly frequency to account for what most of time is just a single day mismatch simply sounds unjustifiable.

I do believe the concept is generally valid. So if we were in a professional, real-money, wide portfolios managing, I would dig deeper on those nuances and devise a way to possibly allow greater rigor than an observation shifted forward.

At the same time, it would remain largely questionable whether the above would have a statistically significant effect with respect to the kind of analyses we are carrying on. It is probably better to spend this work time on advancing the schedule in other directions (models).

#18