Create COMLR data parser
I am aware of the pre-existing parser https://github.com/JustinMShea/ExpectedReturns/blob/master/inst/parsers/Commodities-for-the-Long-Run.R. However, it was added many days ago and would have needed revision. In particular:
- it isn't working as is for me
- it is not directly based on 'rio'
- its file name is potentially troublesome
- it now has a different style from most other parsers

These aspects convinced me to add a new parser, aiming for readability and consistency with code written so far. Since I am not the author of that parser and at this time I am unsure whether it would be allowed, I won't delete it myself.

One interesting aspect is how the two parsers are treating the state variables. I went for classic 1/0 binary because:
- I suppose it's more efficient
- it allows to have an xts obj
- it shouldn't display particular downsides (conversions, etc.)

Of course, can always switch to 'factor' if this indeed turns out to be needed or helpful.

#19