Add multi-horizon investment styles models
LmHorizon(), added with e0ba97d, is now applied on investment styles regressions as well. The function revision is in order to take into account these other models too, allowing in the end a more flexible structure.

This completes the targets to reproduce from the paper. More in depth reviews are sought after and welcome. Especially on measurements, where there are slightly fuzzy boundaries.

#19