Add closing prices vs. settlement prices
As discussed in more depth with @JustinMShea. The 'Last' to 'Settle' edit for equity indexes is really just for computational convenience, as the majority of instruments series in turn have 'Settle' but not e.g. 'Last'. Also, we should use variables consistently throughout so, as long as this is explained in docs & comments, it shouldn't cause confusion.

#22