Improve futures data loading
Still aiming to keep everything quite simple. This commit mainly adds a parameter to the function so to specify which files one wants to load, the main use case I have in mind for this is allowing us to drop files into the pre-selected target folder without having to worry too much unwanted files will be unnecessarily loaded into the environment.

Also, the function is now renamed 'Load*' instead of the former 'Import*' because essentially load()-ing is all it has to achieve.

#22