Add MomSignal() function for trading signals
By and large, for this well structured pieces of work I believe is probably best to adopt the good old script-based development. This modular approach greatly helps and makes things easier to manage, as here we'll probably have a number of packages involved. Last but not least, I think it's also in the direction of having code (much) easier to reuse and test.

Specifically, the function added is complete of minimal docs. At the same time, the package structure gained 'PerformanceAnalytics', which is already happily used for calculations all around and more of its functionality adoption is quite likely forthcoming. However, at the moment, I'm not completely sure 'Imports' is the most appropriate field in our case.

MomSignal() structure has already next signals additions in mind. At the moment, SIGN is the only method implemented and a very first usage example is provided in the corresponding vignette.

#22