Let ExpectedReturns::MomSignal work with xts objs
Conversions to xts are now in place. This is both helpful on its own for our purposes and also because many of the packages in the reference framework either rely or work better with xts objs.

Now the function is more flexible, trying to catch when possible both the more comprehensive data structures inputs we'll likely need anyways (for further purposes) and the efficiency of a package designed to work with time-series (with particular attention to financial applications).

For future reference, the above points make me wondering what is left to consider for a package-wide adoption of 'data.table'. Simplifying a bit, it's a (much) improved 'data.frame' and thus it's intellegible how its data structures would beautifully serve goals from end to end, hopefully once and for all.

#22