Revise SIGN and MA signals
Both had incorrect calculations. In particular, to notice is 'round(lookback/12)' and 'lookback=261' in the vignette, involving the broader time scaling aspect. It will be proposed for discussion in same issue comments.

Switched from 'y' to 's' in values return just to make a bit more evident they're two separate entities, and also because would be immediate to eventually add other computed values too.

'na.rm' probably shouldn't be there at all in the end and, as detailed before, they would be removed as soon as data is filled-forward.

#22