Add SMT momentum signal
TREND and SMT share most calcs, they're squashed together so to DRY. It is especially so when SMT splitting over subperiods is overlooked, as is the case here for now. Of course, this doesn't necessarily imply they end up in positions agreement.

Both aspects shall be better clarified in object or long-format docs.

#22