Resolve most MomSignal() related NOTEs
- coping with <R CMD check --as-cran>, which with always a pinch of surprise only fully supports 'base' by default
- 'vcov.' partial matching

On "no visible binding for global variable", within() seems to be blamed. This should refer to its interactive vs. programming usage. 'utils::globalVariables' or 'rlang::.data' shall resolve the problem, however the flow in question was picked because of its exressiveness and thus if I had to prefix variables with e.g. '.data' I'd perhaps revert to the usual reassignement flow.

#22