Add structure on MOM signals in vignette
In my humble opinion, the main reference paper (which is widely available), already does a great job in illustrating the signals concisely. However, as usual, we should provide definitions as well for completeness. So this commit adds a bit of structure to blend the discussion as we go toward piece of work completion on signals.

The hope is to provide a fairly informative illustration for users, especially with respect to the functionality provided. As a minor aspect, I'm still slightly undecided on whether it's better to split the signals' code chunk to provide step-by-step examples or if placing them together would be easier to deal with in sight of chunks executions.

#22