Add Baltas-Kosowski (2013)
The working paper is made public by the CME Group.

IMHO, it would require a little bit of stretch to say that Baltas-Kosowski (2013) is an updated version of Baltas-Kosowski (2012). Of course, the authors are still studying those phenomena, but after a first glance many aspects do change in several directions in the later work. However, many arguments do not seem mutually exclusive and thus when it makes sense I'd try to have comprehensive implementations.

#22