Add AggregateByType() and exclude conversions
AggregateByType() is another utility function. As with others added its main scope is to help dealing with the nested objects produced, for some tasks.

Conversions to data.frame are unnecessary in all the computations coded so far. As known, dealing with 'xts' is more efficient (and pleasant) in those circustances. However, soon a panel format  will be needed or convenient to have in order to reproduce results and so the code commented out now will be reused for tat purpose.

#22