Add contracts closing prices and YZ-vol plots
This is to reproduce figure 1. Several, more or less evident, differences are present. For prices, sometimes the comparison seems to hold different futures contracts. For volatility, especially when considering percentages, values look diverging from the paper by say a factor of 10 in many instances. The matter is receiving a closer scrutiny.

#22