Introduce ExpectedReturns::SignalStats()
A small function to compute signals statistics. It is directed toward momentum signals here, but same analyses look useful for other signals as well. I've only tested it on ExpectedReturns::MomSignal() output, however I would expect it to hold similarly for other circumstances or anyways the structures aren't extremely difficult to obtain and input. Slight adjustments may occur over time to extend the function and catch other instances.

Correlation and activity matrices examples are added in vignette, in place of the previous snippets from with they originate. Object docs are provided as usual.

One interesting aspect is about internally produced NAs which may potentially cause breaks or logical errors. However, there are several dimensions involved already and - because all the statistics contemplated so far should involve a number of signals > 2 - we may want to introduce consistency checks among them in the first place (as opposed to what happens e.g. on L71). Or, for the time being, doing so in ExpectedReturns::MomSignal() which is nevertheless under review.

#22