Add signals agreement matrix in SignalStats()
The signals postion agreement matrix is in table 3, panel A (second "column" matrix). However, it feels somewhat ambiguous what exactly to do when comparing signals than do not allow for inactivity vs signals that do allow it. In logical terms it seems it still counts as disagreement, but the number is deflated with respect to when comparing against signals that do not allow such inactivity periods by construction.

Also, results are displayed differently with respect to the paper and simply follow from the utils::combn algorithm. Would be nice to display them in a perhaps more friendly manner if it doesn't mean to have to reivent the wheel.

Lastly, took a chance to switch from 'cor' to 'correlation'.

#22