Add futures contracts summary stats 2/
Added normality and autocorrelation tests.

On these and other tests to be added, packages needed are stated explicitly, however it's both questionable whether users would like to install entire packages for a test statistic they may be  not interested about and a pity to fail vignette knitting for a missing package. Wondering if it's better to try-catch in such instances?

#22