Add futures contracts summary stats 4/
- "fixed" annualized average returns and volatility
- introduced Symmetric Downside-Risk Sharpe Ratio (DRSR)

As of the present commit, values in table 1 don't appear to be resembled within a reasonable margin. It feels somewhat weird because commonly used and tested functions are being used, I'm wondering what's wrong.

For example, couldn't find how which returns they are working with: I suspect simple returns by looking at their DRSR definition, however it seems there's a prominent influence on volatility.

Also, the DRSR is referenced to be annualized, but scaling the ratio resulted in considerable deviations.

#22