Add TSMOM strategy 1/
Initial commit for the time series momentum function, to be further revised and expanded. In particular, the 'holding' parameter is implemented but not currently used.

Also, TSMOM() may be useful as an exported function for everyone to use elsewhere with ease. I've refrained from doing so right now because it in turn relies on EstimateVolatility(), which would then need to be exported as well. Furthermore requiring to add 'TTR' as an 'ExpectedReturns' dependency.

#22