Add TSMOM strategy 2/
Replicating table 4 appears to be a 4D problem. Added 4/6 of the TSMOM strategies performance measurements via TsmomStats().

Observations:
- as noted in 266a799, 'holding' is not currently used and this is why values appear to be repeated in the output at the time of writing;
- as noted in a 266a799 TODO, it seems volatility estimation happens on a daily basis, but these analyses are explicitly reported to be on monthly data. In this perspective, n=2 passed-though to TTR::volatility() is likely incorrect. Trying to better figure this aspect in conjunction with other dimensions involved.

Perhaps it would be useful to export this function as well, so useRs may call it in their console for interactive analysis or use it for reports without having to dig in the vignette and copy-paste elsewhere.

#22