Revise MomSignal() 'signal' passthru
Chiefly because calling MomSignal() from within TSMOM() - and a fortiori TsmomStats() - caused 'frames' misbehaviors (may be related to the fact the latter two are defined in '.GlobalEnv'). Instead of begin playing with environments (which given the vignette developement flow may be not so portable), replaced the call storage that was only used for one variable. Minor fixes on variables names were done as well.

However, as per the present commit, TsmomStats() still fail when called on all signals. This appears to be due to 'TREND' OR 'SMT', specifically for the lagging used within sandwhich::NeweyWest, and it's receiving closer scrutiny.

#22