Add TSMOM strategy 3/
Added Table 5 replication code in Table 4 replication flow. It appears convenient to do so to DRY. Even if we would like to disentangle the two tables for illustration purposes later on.

This being said, it's an initial commit and numbers seem off with respect to the working paper right now, likely reflecting several other aspects under review that will be fixed.

#22