Add TSMOM predictability 6/
Code to reproduce Figure 5 in Baltas-Kosowski (2013), not the 2012 version. In the latter case I'm not sure exactly what they did and while looking at the former for further details I found that only part of it was kept in analyses. Thus decided to try reproducing the former.

The TREND mom signal is commented out because it produces an error. I think this is the same error highlighted in previous commits messages and has to do with the time handling aspect, which will hopefully receive a specific thoroughly review.

#22