Add data section in vignette
Always a pleasure to see a well curated website sharing open source academic research financial data.

Data sets seem to be detailed in a similar fashion to French's library, they're smaller in number as likely "standard" factors will still be from the latter.

Because this library is updated regularly, I'm wondering if a quick extension of ExpectedReturns::GetFactors() may be worth it to always attempt downloading fresh data.
Likewise, since the library is open sourced, I'm guessing we can share R data objects as we're doing with some French's data. Of course, always with proper credit to the respective authors. See #16.

#23