Extend ExpectedReturns::GetFactors() to q-factors
If the function feels artisanal is because it is. It's simply a convenience function to download data files we need from academic research websites, but there isn't anything to connect to systematically and all the data sets from different source have their peculiarities to account for (which also motivates why sources are being left as separate as possible, even at the cost of some code repetition).

'rio' has been introduced as a dependency because it simplifies the web import (and as we're using it for data parsers). If approved, then it might be a good idea to change the src='FF' flow accordingly. Although there must be some sort of internal binary conversion, I haven't checked yet what actually happens to the original file (should be erased).

#23