Direct FF files downloads to tempdir
Workaround to avoid manual deletion of FF's files from package root, as 'tempdir' should be cleaned up whenever the R session is closed.

As mentioned in 0522589, if 'rio' has to be imported nevertheless then it makes sense to use it here as well. Haven't done that yet because, as a consequence, all the parsing done around French's files will need to be revised according to how 'rio' will parse the file.