Edit minor references details
"Digesting Anomalies: An Investment Approach" is often reported on OUP files to have been published on 2014, sometimes they refer to 2015 (maybe for when it was made available online?). Anyways authors themeselves always report the work to be on 2015, so using this date to avoid confusion with respect to works we reference to.

#23