Create Daniel-Hirshleifer-Sun (2020) data parser
The replication data set is made publicly available by one of the authors, Prof. Kent D. Daniel. To my knowledge so far, it doesn't seem that authors plan to keep their factors data sets updated. For this reason a standalone parser has been coded, as opposed to potential fresh data downloads attempts.

To note is that their 3-factor model (what they call "BF3") includes a market proxy factor which is not provided in the downloadable data set. Beacuse they refer to the CAPM and the FF model, I assumed it's the same one and thus added it to the original data set in order to allow ourselves & users to download the whole thing ready for analyses (methodological choices on breakpoints different from the standard approach should only involve PEAD and FIN).

The data set is needed in #23.

See also #16