Upload Daniel-Hirshleifer-Sun (2020) data set
The replication data set is made publicly available by one of the authors, Prof. Kent D. Daniel. To the best of my knowledge, it seems to have been shared without particular restrictions. Of course, all credits for data remain to the respective authors, whose work is cited.

Again, to note is that their 3-factor model (what they call "BF3") includes a market proxy factor which is not provided in the downloadable data set. Beacuse they refer to the CAPM and the FF model, I assume it's the same one and thus added it to the original data set in order to allow ourselves & users to download the whole thing ready for analyses (methodological choices on breakpoints different from the standard approach should only involve PEAD and FIN).

Documentation has been added, at least in order to provide a brief introduction. Their factors construction has been stressed (as in previous comparable instances) for future purposes.

The data set is needed in #23.

See also #16