Add q-factor vs DHS (2020) spanning regressions
Initial commit to replicate table 3 via SpanningRegressions().

There are more or less slight deviations from "Which Factors?", both of them are related to the fact Hou-Mo-Xue-Zhang (2019) are actually constructing factors whereas we are not (something I wish for the future of our package). In particular:
- Hou-Mo-Xue-Zhang (2019) construct the Daniel-Hirshleifer-Sun (2020) model on the 1967-01 to 2016-12 sample period. We are using Daniel-Hirshleifer-Sun (2020) own kindly provided series - which at this point have been likely disseminated only recently. Thus we are constrained to consider the 1972-07 to 2018-12 period at most. In this context, we can clearly always chop the series on the upper end, but not augment it backwards.
- Hou-Mo-Xue-Zhang (2019) raise some concerns on Daniel-Hirshleifer-Sun (2020)'s factors construction, especially when it deviates from the traditional approach. For this reason they construct and provide results for the Daniel-Hirshleifer-Sun (2020) obtained with the standard approach. Again, at this time we are not constructing factors and thus we are unable to include such additional set of results.

As in previous comparable instances, 'MKT.RF.1' (Hou-Xue-Zhang's market factor) regressions are included mostly to avoid confusing variables names, but also because it slightly differs from French's 'MKT.RF' factor (should be beacuse of 'RF'). Anyways, we may easily exclude these additional regressions (not included in the paper).

#23