Add Barillas-Shanken (2018) data set in vignette
Although no replication data set is readily available to my knowledge, it seems we already have all we need. In particular, it's great we already worked on Asness-Frazzini (2013) and thus can easily obtain those factors (even if its data parser maybe needs some edits, it does the job).

In order to (a) avoid confusion in the vignette and focus on the main analytical flow instead, and (b) be able to offer a better documentation for such data set, this code may found a better place as a standalone parser. However, it remains unclear wether we would ever be able to redistribute the entire data set because one of the factors (HML.DEV) comes from AQR's data sets.

#23