Add q-factors data set references in vignette
To reference the more comprehensive docs on these factors. Although we may add some brief summaries in the vignette as well, I think that's where most users would prefer to access this information (more closely related to what they download and can work on in practice, and possibly easier to grab). Otherwise there are the original references, that of course I recommend reading anyways.

#23